Sincronização das assinaturas apenas via cron para montar o grid no admin do Magento.
