+++ etc/crontab.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" ?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Cron:etc/crontab.xsd">
+    <group id="default">
+        <job instance="Vindi\Payment\Cron\SyncSubscriptions" method="execute" name="vindi_payment_sync_subscriptions">
+            <schedule>0 * * * *</schedule>
+        </job>
+    </group>
+</config>
